+++ lib/src/revset.rs
@@ -457,6 +457,20 @@ impl<St: ExpressionState> RevsetExpression<St> {
         })
     }
 
+    /// First-parent ancestors of `self`, including `self`.
+    pub fn first_ancestors(self: &Rc<Self>) -> Rc<Self> {
+        self.first_ancestors_range(GENERATION_RANGE_FULL)
+    }
+
+    /// First-parent ancestors of `self` in the given range.
+    pub fn first_ancestors_range(self: &Rc<Self>, generation_range: Range<u64>) -> Rc<Self> {
+        Rc::new(Self::Ancestors {
+            heads: self.clone(),
+            generation: generation_range,
+            parents_range: 0..1,
+        })
+    }
+
     /// Children of `self`.
     pub fn children(self: &Rc<Self>) -> Rc<Self> {
         self.descendants_at(1)
@@ -745,6 +759,17 @@ static BUILTIN_FUNCTION_MAP: Lazy<HashMap<&'static str, RevsetFunction>> = Lazy:
             GENERATION_RANGE_FULL
         };
         Ok(roots.descendants_range(generation))
+    });
+    map.insert("first_ancestors", |diagnostics, function, context| {
+        let ([heads_arg], [depth_opt_arg]) = function.expect_arguments()?;
+        let heads = lower_expression(diagnostics, heads_arg, context)?;
+        let generation = if let Some(depth_arg) = depth_opt_arg {
+            let depth = expect_literal("integer", depth_arg)?;
+            0..depth
+        } else {
+            GENERATION_RANGE_FULL
+        };
+        Ok(heads.first_ancestors_range(generation))
     });
     map.insert("connected", |diagnostics, function, context| {
         let [arg] = function.expect_exact_arguments()?;